% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caesar.R
\name{caesar}
\alias{caesar}
\title{Encrypt and Decrypt text using the Caesar cipher}
\usage{
caesar(text, distance = 3, reverse = FALSE)
}
\arguments{
\item{text}{String to be ciphered or deciphered}

\item{distance}{How far to move the characters in the direction you choose.}

\item{reverse}{If TRUE, deciphers the coded text}

\item{direction}{Move the original characters the right or to the left}
}
\value{
String of the ciphered/deciphered text
}
\description{
Encrypt and Decrypt text using the Caesar cipher
}
\examples{
# Please see this for more info.
# https://en.wikipedia.org/wiki/Caesar_cipher

caesar("HELLO WORLD")
caesar("khoor zruog", reverse = TRUE)

caesar("HELLO WORLD", distance = 12)
caesar("tqxxa iadxp", distance = 12, reverse = TRUE)

}
