% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caesar.R
\name{seed_cipher}
\alias{seed_cipher}
\title{Encrypt and decrypt text using pseduorandom number generation based on the seed set.}
\usage{
seed_cipher(text, seed = 64, decrypt = FALSE)
}
\arguments{
\item{text}{String to be ciphered or deciphered}

\item{seed}{A single number to set the seed which will pseudorandomly rearrange
the original characters}

\item{decrypt}{If TRUE, deciphers coded text}
}
\value{
String of the ciphered/deciphered text
}
\description{
Encrypt and decrypt text using pseduorandom number generation based on the seed set.
}
\examples{
seed_cipher("As a rule, men worry more about what they can't see than about what they can.")
seed_cipher("O8GdG@SfhEG9hUG*A@nG9A@hGd:AS_G*Wd_G_WhnG)dUg_G8hhG_WdUGd:AS_G*Wd_G_WhnG)dUx", decrypt = TRUE)


seed_cipher("Men willingly believe what they wish.", seed = 2354)
seed_cipher("q39l*D66D9;6.l\%36D3d3l*<p4l4<3.l*D <h", seed = 2354, decrypt = TRUE)

seed_cipher("What we wish, we readily believe, and what we ourselves think, we imagine others think also.", seed = -100)
seed_cipher("}Nf,&3C&3etN4&3C&vCf>eB8&xCBeC]C4&fj>&3Nf,&3C&P9vtCB]Ct&,Nej:4&3C&e-fZejC&P,NCvt&,Nej:&fBtP ", seed = -100, decrypt = TRUE)
}
