% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caesar.R
\name{caesar}
\alias{caesar}
\title{Encrypt and decrypt text using the Caesar cipher.}
\usage{
caesar(text, distance = 3, decrypt = FALSE)
}
\arguments{
\item{text}{String or vector of strings to be ciphered or deciphered.}

\item{distance}{A single whole number for how far to move the characters in the direction (positive or negative) you choose. If not a whole number, it will be rounded to nearest whole number.}

\item{decrypt}{If TRUE, (not default) deciphers the coded text.}
}
\value{
String of the ciphered/deciphered text
}
\description{
Encrypt and decrypt text using the Caesar cipher.
}
\examples{
# Please see this for more info.
# https://en.wikipedia.org/wiki/Caesar_cipher

caesar("Experience is the teacher of all things.")
caesar("HAshulhqfhclvcwkhcwhdfkhucricdoocwklqjva", decrypt = TRUE)

caesar("Veni, vidi, vici.",  distance = 40)
caesar(",S1WKN9WRWKN9WQWL", distance = 40, decrypt = TRUE)

caesar("No one is so brave that he is not disturbed by something unexpected.", distance = -12)
caesar("Bc[cb:[,g[gc[{f]j:[h>]h[>:[,g[bch[;,ghif{:;[{m[gca:h>,b<[ib:ld:}h:;`",
       distance = -12, decrypt = TRUE)
}
