% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caesar.R
\name{caesar}
\alias{caesar}
\title{Encrypt and Decrypt text using the Caesar cipher}
\usage{
caesar(text, direction = "left", distance = 3, reverse = FALSE)
}
\arguments{
\item{text}{String to be ciphered or deciphered}

\item{direction}{Move the original characters the right or to the left}

\item{distance}{How far to move the characters in the direction you choose.}

\item{reverse}{If TRUE, deciphers the coded text}
}
\value{
String of the ciphered/deciphered text
}
\description{
Encrypt and Decrypt text using the Caesar cipher
}
\examples{
# Please see this for more info.
# https://en.wikipedia.org/wiki/Caesar_cipher

caesar("cats are friends!")
caesar("fdwv@duh@iulhqgv&", reverse = TRUE)

caesar("cats are friends!", direction = "right", distance = 40)
caesar("ki10:izm:nzqmvl0;", direction = "right", distance = 40, reverse = TRUE)

caesar("cats are #1 friend!", direction = "left", distance = -12)
caesar(", hgy f&ypy\%f:&b@z", direction = "left", distance = -12, reverse = TRUE)
}

