% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caesar.R
\name{seed_cipher}
\alias{seed_cipher}
\title{Encrypt and Decrypt text using pseduorandom number generation}
\usage{
seed_cipher(text, seed = 64, reverse = FALSE)
}
\arguments{
\item{text}{String to be ciphered or deciphered}

\item{seed}{A number to set the seed which will pseudorandomly rearrange
the original characters}

\item{reverse}{If TRUE, deciphers coded text}
}
\value{
String of the ciphered/deciphered text
}
\description{
Encrypt and Decrypt text using pseduorandom number generation
}
\examples{
seed_cipher("cats are friends!")
seed_cipher("bc204c5d495ud?:08", reverse = TRUE)


seed_cipher("cats are friends!", seed = 2354)
seed_cipher("li\%1sid3szdp3 j1g", seed = 2354, reverse = TRUE)

seed_cipher("cats are #1 friend!", seed = -100)
seed_cipher("hd2imdjpmfm-jcpe&q", seed = -100, reverse = TRUE)
}

